Fix load callback spam part 4
It looks like the core issue was that the load future was
not being completed when ticket level fell. However, by
adding code to complete it, it caused the NOISE status
to start executing on the same chunks in parallel.
So, no longer complete the future when ticket level falls.
Instead, add code directly into the FULL future callbacks
to CHECK the ticket level / load status BEFORE scheduling.
This should prevent double scheduling problems.